Robust distinction between IDENTIFIER and TYPE_NAME tokens.
To distinguish the tokens we used to track the lexer context,
which is fragile. This patch implements a better way -
combine the two tokens into a common parser rule and let
the parser context decide if the declaration is valid,
which it already does by checking the naming conflicts.

TEST=WebGL conformance tests.

Review URL: https://codereview.appspot.com/8797044

git-svn-id: http://angleproject.googlecode.com/svn/trunk@2190 736b8ea6-26fd-11df-bfd4-992fa37f6226